Move current analyzer tests to test/server/ folder
