+++ packages/labs/analyzer/package.json
@@ -17,7 +17,8 @@
   "type": "module",
   "scripts": {
     "build": "wireit",
-    "test": "wireit"
+    "test": "wireit",
+    "test:server": "wireit"
   },
   "wireit": {
     "build": {
@@ -36,8 +37,13 @@
       "clean": "if-file-deleted"
     },
     "test": {
+      "dependencies": [
+        "test:server"
+      ]
+    },
+    "test:server": {
       "#comment": "The quotes around the file regex must be double quotes on windows!",
-      "command": "cross-env NODE_OPTIONS=--enable-source-maps uvu test \"_test\\.js$\"",
+      "command": "cross-env NODE_OPTIONS=--enable-source-maps uvu test/server \"_test\\.js$\"",
       "dependencies": [
         "build",
         "../../lit:build"
