+++ packages/labs/analyzer/src/test/server/javascript/classes_test.ts
