+++ packages/labs/analyzer/src/test/server/javascript/functions_test.ts
